Fixed layout of editor and organization of react components (#20)
