+++ src/custom/Footer.js
@@ -1,0 +1,6 @@
+const Footer = () => {
+
+    return <></>
+}
+
+export default Footer;
