+++ src/custom/Header.js
@@ -1,0 +1,14 @@
+const Header = () => {
+
+    return <>
+        <div style={{float: "right", marginRight: "5px"}}>
+            <a href={"https://github.com/KNowledgeOnWebScale/dataflow-visualization#readme"} target="_blank"
+               rel="noreferrer"
+            >
+                Read the docs
+            </a>
+        </div>
+    </>
+}
+
+export default Header;
