+++ src/data/examples.js
@@ -1,0 +1,38 @@
+import {
+    edgesJSON as edgesJSON1,
+    globalDefaultsJSON as globalDefaultsJSON1,
+    nodesJSON as nodesJSON1
+} from "./examples/exampleData1";
+import {
+    edgesJSON as edgesJSON2,
+    globalDefaultsJSON as globalDefaultsJSON2,
+    nodesJSON as nodesJSON2
+} from "./examples/exampleData2";
+import {
+    edgesJSON as edgesJSON3,
+    globalDefaultsJSON as globalDefaultsJSON3,
+    nodesJSON as nodesJSON3
+} from "./examples/exampleData3";
+import {
+    edgesJSON as edgesJSON4,
+    globalDefaultsJSON as globalDefaultsJSON4,
+    nodesJSON as nodesJSON4
+} from "./examples/exampleData4";
+import {
+    edgesJSON as edgesJSON5,
+    globalDefaultsJSON as globalDefaultsJSON5,
+    nodesJSON as nodesJSON5
+} from "./examples/exampleData5";
+import {
+    edgesJSON as edgesJSON6,
+    globalDefaultsJSON as globalDefaultsJSON6,
+    nodesJSON as nodesJSON6
+} from "./examples/exampleData6";
+
+
+export const examples = [
+    [globalDefaultsJSON1, nodesJSON1, edgesJSON1], [globalDefaultsJSON2, nodesJSON2, edgesJSON2],
+    [globalDefaultsJSON3, nodesJSON3, edgesJSON3], [globalDefaultsJSON4, nodesJSON4, edgesJSON4],
+    [globalDefaultsJSON5, nodesJSON5, edgesJSON5], [globalDefaultsJSON6, nodesJSON6, edgesJSON6],
+];
+
